refactor: ColorTapPlayArea and EmojiMathPlayArea use game hook (#291)
Closes #291
